requirements.txt
beautifulsoup4==4.14.2
lxml==6.0.2
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2

# Dependencias para LayoutLMv3
transformers
torch
torchaudio
torchvision
timm==1.0.7
Pillow
pytesseract==0.3.10
